chore(lint): mypy --strict — add type annotations
- Parametrise generic types: dict[str, Any], asyncio.Task[Any], TypeDecorator[uuid.UUID]
- Add return-type annotations to _dispatch, _run, _lookup_key_by_name, _generate
- Type _patch_env_value's generator parameter as Callable[[], str]
- Cast .json() returns to dict[str, Any] / Optional[str] to satisfy no-any-return
- Cast bool from comparison and int from math.ceil
- Drop unused 'type: ignore' comments and replace [import] with [import-not-found]
- Add HTTPException 404 guards before .attribute access on session.get(...) results
- Widen proxy_messages return type to include Response
- Add 'date' import to models.py (was missing)
- Add 'admin_key is None' guard via narrowing on prior usage

mypy --strict src/tourniquet/ now returns 0 errors.

+++ src/tourniquet/billing/suggestions.py
@@ -237,4 +237,4 @@ async def get_rolling_avg_from_db(
     if not rows:
         return 0
     total = sum(row[1] for row in rows)
-    return math.ceil(total / days)
+    return int(math.ceil(total / days))
